requirements.txt
Pillow==10.1.0
pypdf2==3.0.1
typing-extensions==4.9.0
